QuizFeedbackEngine/pyproject.toml
[tool.poetry]
name = "quiz-feedback-engine"
version = "0.1.0"
description = "LangGraph-powered quiz feedback system with LLM integration"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0.0"
fastapi = ">=0.121.0"
langchain = ">=1.0.5"
langchain-openai = ">=1.0.2"
langgraph = ">=1.0.2"
pydantic = ">=2.12.4"
uvicorn = ">=0.38.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^1.0.0"
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
